requirements.txt
Flask==2.3.3
requests==2.31.0
debugpy # Required for debugging.
google-genai==1.2.0
google-cloud-firestore==2.19.0
